Add files via upload
Updated solution file to hopefully resolve issues with UserRoleRelation & UserRoleActivation. Also created WIP SystemManager class (separated). The driver has been slightly modified to be in line with SystemManager functionality, though this also needs to be finished. As far as I can tell there are no issues with this version and it does run correctly.